Move LinekdMerkleTree test to own file.
