+++ packages/common/test/trees/MerkleTree.test.ts
@@ -1,123 +1,106 @@
 import { beforeEach } from "@jest/globals";
-import { Field, Poseidon } from "o1js";
+import { Field } from "o1js";
 
-import {
-  createLinkedMerkleTree,
-  InMemoryLinkedMerkleTreeStorage,
-  log,
-} from "../../src";
+import { createMerkleTree, InMemoryMerkleTreeStorage, log } from "../../src";
 
-describe.each([4, 16, 254])("cachedMerkleTree - %s", (height) => {
-  class LinkedMerkleTree extends createLinkedMerkleTree(height) {}
+describe.each([4, 16, 256])("cachedMerkleTree - %s", (height) => {
+  class RollupMerkleTree extends createMerkleTree(height) {}
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  class RollupMerkleTreeWitness extends RollupMerkleTree.WITNESS {}
 
-  let store: InMemoryLinkedMerkleTreeStorage;
-  let tree: LinkedMerkleTree;
+  let store: InMemoryMerkleTreeStorage;
+  let tree: RollupMerkleTree;
 
   beforeEach(() => {
     log.setLevel("INFO");
 
-    store = new InMemoryLinkedMerkleTreeStorage();
-    tree = new LinkedMerkleTree(store);
+    store = new InMemoryMerkleTreeStorage();
+    tree = new RollupMerkleTree(store);
   });
 
   it("should have the same root when empty", () => {
     expect.assertions(1);
 
     expect(tree.getRoot().toBigInt()).toStrictEqual(
-      LinkedMerkleTree.EMPTY_ROOT
+      RollupMerkleTree.EMPTY_ROOT
     );
   });
 
   it("should have a different root when not empty", () => {
     expect.assertions(1);
 
-    tree.setLeaf(1n, 1n);
+    tree.setLeaf(1n, Field(1));
 
     expect(tree.getRoot().toBigInt()).not.toStrictEqual(
-      LinkedMerkleTree.EMPTY_ROOT
+      RollupMerkleTree.EMPTY_ROOT
     );
+  });
+
+  it("should have the same root after adding and removing item", () => {
+    expect.assertions(1);
+
+    tree.setLeaf(1n, Field(1));
+
+    const root = tree.getRoot();
+
+    tree.setLeaf(5n, Field(5));
+    tree.setLeaf(5n, Field(0));
+
+    expect(tree.getRoot().toBigInt()).toStrictEqual(root.toBigInt());
   });
 
   it("should provide correct witnesses", () => {
     expect.assertions(1);
 
-    tree.setLeaf(1n, 1n);
-    tree.setLeaf(5n, 5n);
+    tree.setLeaf(1n, Field(1));
+    tree.setLeaf(5n, Field(5));
 
-    const witness = tree.getWitness(5n).leafCurrent;
+    const witness = tree.getWitness(5n);
 
-    expect(
-      witness.merkleWitness
-        .calculateRoot(
-          Poseidon.hash([
-            witness.leaf.value,
-            witness.leaf.path,
-            witness.leaf.nextPath,
-          ])
-        )
-        .toBigInt()
-    ).toStrictEqual(tree.getRoot().toBigInt());
+    expect(witness.calculateRoot(Field(5)).toBigInt()).toStrictEqual(
+      tree.getRoot().toBigInt()
+    );
   });
 
   it("should have invalid witnesses with wrong values", () => {
     expect.assertions(1);
 
-    tree.setLeaf(1n, 1n);
-    tree.setLeaf(5n, 5n);
+    tree.setLeaf(1n, Field(1));
+    tree.setLeaf(5n, Field(5));
 
     const witness = tree.getWitness(5n);
 
-    expect(
-      witness.leafCurrent.merkleWitness.calculateRoot(Field(6)).toBigInt()
-    ).not.toStrictEqual(tree.getRoot().toBigInt());
+    expect(witness.calculateRoot(Field(6)).toBigInt()).not.toStrictEqual(
+      tree.getRoot().toBigInt()
+    );
   });
 
   it("should have valid witnesses with changed value on the same leafs", () => {
     expect.assertions(1);
 
-    tree.setLeaf(1n, 1n);
-    tree.setLeaf(5n, 5n);
+    tree.setLeaf(1n, Field(1));
+    tree.setLeaf(5n, Field(5));
 
-    const witness = tree.getWitness(5n).leafCurrent;
+    const witness = tree.getWitness(5n);
 
-    tree.setLeaf(5n, 10n);
+    tree.setLeaf(5n, Field(10));
 
-    expect(
-      witness.merkleWitness
-        .calculateRoot(
-          Poseidon.hash([Field(10), witness.leaf.path, witness.leaf.nextPath])
-        )
-        .toBigInt()
-    ).toStrictEqual(tree.getRoot().toBigInt());
+    expect(witness.calculateRoot(Field(10)).toBigInt()).toStrictEqual(
+      tree.getRoot().toBigInt()
+    );
   });
 
-  it("should return zeroNode", () => {
-    expect.assertions(3);
-    const MAX_FIELD_VALUE: bigint = BigInt(2 ** 53 - 1);
-    const zeroLeaf = tree.getLeaf(0n);
-    expect(zeroLeaf.value.toBigInt()).toStrictEqual(0n);
-    expect(zeroLeaf.path.toBigInt()).toStrictEqual(0n);
-    expect(zeroLeaf.nextPath.toBigInt()).toStrictEqual(MAX_FIELD_VALUE);
-  });
+  it("should throw for invalid index", () => {
+    expect.assertions(2);
 
-  it("should return zeroNode", () => {});
-});
+    const index = 2n ** BigInt(height) + 1n;
 
-// Separate describe here since we only want small trees for this test.
-describe("Error check", () => {
-  class LinkedMerkleTree extends createLinkedMerkleTree(4) {}
-  let store: InMemoryLinkedMerkleTreeStorage;
-  let tree: LinkedMerkleTree;
-
-  it("throw for invalid index", () => {
-    log.setLevel("INFO");
-
-    store = new InMemoryLinkedMerkleTreeStorage();
-    tree = new LinkedMerkleTree(store);
     expect(() => {
-      for (let i = 0; i < 2n ** BigInt(4) + 1n; i++) {
-        tree.setLeaf(BigInt(i), 2n);
-      }
+      tree.setLeaf(index, Field(1));
+    }).toThrow("Index greater than maximum leaf number");
+
+    expect(() => {
+      tree.getNode(0, index);
     }).toThrow("Index greater than maximum leaf number");
   });
 });
